move partial dates to hook for readability
