+++ packages/dnb-eufemia/src/components/date-picker/hooks/usePartialDates.ts
@@ -1,0 +1,23 @@
+import { useRef } from 'react'
+
+export type PartialDates = {
+  partialDate?: string
+  partialStartDate?: string
+  partialEndDate?: string
+}
+
+export default function usePartialDates() {
+  const partialDatesRef = useRef<PartialDates>({
+    partialStartDate: null,
+    partialEndDate: null,
+  })
+
+  function setPartialDates(partialDates: PartialDates) {
+    partialDatesRef.current = {
+      ...partialDatesRef.current,
+      ...partialDates,
+    }
+  }
+
+  return { partialDatesRef, setPartialDates }
+}
